Add a new driver for FYGen.
I copied files from:
  https://github.com/mattwach/fygen

then created a wrapper to allow them to be used.

This should, in theory, give better bode plot performance
because fygen is a closed-loop system (i.e. instead
of sleeping for 0.5 seconds, it implements reads and
retries which will generally allow for < 0.5 second
update times)

All of this still needs to be tested and benchmarked.  Not
yet ready for use.